backend/requirements.txt
# Core Flask
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database (SQLite - no extra driver needed for development)
# psycopg2-binary==2.9.7  # Commented out - causes Windows issues

# Development Tools
python-dotenv==1.0.0
marshmallow==3.20.1

# Testing
pytest==7.4.2
pytest-flask==1.2.0

# Code Quality
black==23.7.0
flake8==6.0.0